NIFI-9666: Removed jackson.version and fixed utilization diagnostic calculations

+++ nifi-framework-api/src/main/java/org/apache/nifi/diagnostics/SystemDiagnostics.java
@@ -168,7 +168,7 @@ public class SystemDiagnostics implements Cloneable {
         if (maxHeap == -1) {
             return -1;
         } else {
-            return Math.round((usedHeap / maxHeap) * 100);
+            return Math.round(((float)usedHeap / maxHeap) * 100);
         }
     }
 
@@ -176,7 +176,7 @@ public class SystemDiagnostics implements Cloneable {
         if (maxNonHeap == -1) {
             return -1;
         } else {
-            return Math.round((usedNonHeap / maxNonHeap) * 100);
+            return Math.round(((float)usedNonHeap / maxNonHeap) * 100);
         }
     }
